style: incia re-estilização dos detalhes da receita
