+++ package.json
@@ -3,12 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@popperjs/core": "^2.11.6",
     "@sheerun/mutationobserver-shim": "^0.3.3",
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
     "axios": "1.1.3",
-    "bootstrap": "5.2.2",
+    "bootstrap": "^5.2.2",
     "clipboard-copy": "4.0.1",
     "history": "5.3.0",
     "lodash": "^4.17.21",
