+++ src/components/Footer/_footer.scss
@@ -7,8 +7,9 @@
   display: flex;
   height: 55px;
   justify-content: space-around;
+  max-width: 1366px;
   position: fixed;
-  width: 100vw;
+  width: 100%;
   z-index: 1;
 
   a:nth-child(1) img {
