+++ src/components/Header/_header.scss
@@ -9,7 +9,7 @@
   justify-content: space-between;
   padding: 0 1rem;
   position: relative;
-  width: 100vw;
+  width: 100%;
   z-index: 1;
 
   .profile-logo {
