+++ src/components/Loading/_loading.scss
@@ -8,7 +8,7 @@
   left: 0;
   position: fixed;
   top: 0;
-  width: 100vw;
+  width: 100%;
   z-index: 0;
 
   .logo-container {
