+++ src/components/_global.scss
@@ -7,3 +7,5 @@
 @import './Header/header';
 
 @import './Footer/footer';
+
+@import './FavoriteAndShare/favoriteAndShare';
