+++ src/pages/styles/_global.scss
@@ -1,1 +1,3 @@
 @import './recipes';
+
+@import './detailsRecipe';
