+++ src/pages/styles/_recipes.scss
@@ -8,7 +8,7 @@
   margin: 0 auto;
   max-width: 375px;
   padding: 1rem 0;
-  width: 100vw;
+  width: 100%;
 
   button {
 
@@ -26,5 +26,5 @@
   justify-content: center;
   overflow-y: auto;
   padding-bottom: 70px;
-  width: 100vw;
+  width: 100%;
 }
