+++ src/pages/styles/favoriteRecipesCard.css
@@ -1,3 +1,4 @@
+/* stylelint-disable selector-class-pattern */
 .container-card {
   background-color: #fed;
   border-radius: 5px;
