+++ src/pages/styles/profile.css
@@ -1,3 +1,4 @@
+/* stylelint-disable selector-class-pattern */
 .perfilPage {
   height: 100vh;
   width: 100vw;
